Add support for Win32 (#11)
* Fix circleci build
* Add github actions build
* Split CMakeLists.txt into separate subdirectories
* Remove declaration of std::hash, using <functional> header
* Add operator{==,!=} for comparing At<T> with T
* Fix infinite loop in WriteVarIntLoop (cast to u8)
* Use data accessor instead of begin()/end() to get char*
* Some integer cast fixes
* Add text::Write{Nat,Int,Float}, to fix overloading issues
* Use CTest for running each unittest executable
* Only include -lstdc++fs if necessary
* Copy gdtoa library into source tree; modify to compile w/ MSVC
* Fix warning when using -2147483648 literal
* Use gtest_force_shared_crt to prevent CRT mismatch

+++ src/text/read.cc
@@ -809,13 +809,13 @@ auto ReadTable(Tokenizer& tokenizer, Context& context) -> OptAt<Table> {
     if (context.features.bulk_memory_enabled() && IsExpression(tokenizer)) {
       // Element expression list.
       WASP_TRY_READ(expressions, ReadElementExpressionList(tokenizer, context));
-      size = expressions.size();
+      size = static_cast<u32>(expressions.size());
       elements =
           ElementList{ElementListWithExpressions{*elemtype_opt, expressions}};
     } else {
       // Element var list.
       WASP_TRY_READ(vars, ReadVarList(tokenizer, context));
-      size = vars.size();
+      size = static_cast<u32>(vars.size());
       elements = ElementList{ElementListWithVars{ExternalKind::Function, vars}};
     }
